replace 2.24 with 2.25
binutils 2.24 causes a compilation error("../bfd/bfd.h:304 error: right-hand operand of comma expression has no effect [-Werror=unused-value]") with gcc 5.2.1 on ubuntu 15.10 while 2.25 does not have this problem.
It is not convenient to specify the gcc 4.7 or other older version with "gclient sync" or manually. Modifying the download_binutils.sh locally to replace 2.24 with 2.25 cannot solve the problem, either, because "gclient sync" always tries to pull this repository version.